+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
-## Unreleased
+## [1.0.0-beta.2] - 2018-08-25
 
 ### Added
 - Can now use Eloquent query builders as resource relationships using the `queriesOne` or `queriesMany`